Set defaults and made classifier function callable
I set the defaults of the arguments for optimized_txt_guid_i2i so that
I can call the script with next to no arguments. This should make it
easier to execute it from Python.
I rearranged the classifier file so that I can more easily call a
function directly instead of trying to execute the whole file from
another python script.