Modified the main script to be a module
Moved the processing of a directory into a function. Moved the
argument parser to its own function. Seperated out the prompt parsing
to its own function. The models are not built in their own
function. The main function has as little as possible in it. The
keyword arguments on the process_source_img_dir function all have the
default values that we will use for the rest of the project.