Modified the classifier code to just run from a better script
Added the imagenet_classes.txt instead of downloading it because I was
having SSL isssues and didn't feel like debugging it. I modified the
classifer.py so that it'll return the files and labels as a list of
dicts instead of having it write it out to a CSV. I added all the
dependencies to environment.yaml because for some reason we only felt
like listing them in the README. Then I just split some imports onto
single lines because PEP told me to.